Convert program from application to library
This was done so that I can write programs using these algorithms
without the need for piston_window and conrod to be running. In my case
I wrote a testing utility that would repeated evolve the creatures only
using the terminal, no graphical output it shown, just the fitness and
time taken to generate.